adapt to sled

+++ package-lists/prepare_patterns.sh
@@ -9,6 +9,7 @@ export EXPLICIT_UNAME=$2
 for i in data/*; do sh $RPM_SOURCE_DIR/preprocess $i; done | perl $RPM_SOURCE_DIR/create-suggests | \
    uniq > data/REST-DVD-SUGGESTS
 for i in gnome kde3 kde4; do
+  if test ! -f toinstall/rest_cd_$i/requires; then continue; fi
   sh $RPM_SOURCE_DIR/preprocess toinstall/rest_cd_$i/requires > t && \
     mv t toinstall/rest_cd_$i/requires
 done
@@ -18,6 +19,11 @@ for i in `cd toinstall && ls -1d *`; do
    mkdir -p utf8_description/$i/ 
    ( echo "+Des:"; echo "nada" ; echo "-Des:" ) >  utf8_description/$i/default
 done
+cat toinstall/*/role | sort | while read role; do
+   mkdir -p utf8_roles/"$role"/
+   echo "=Cat: Nada" > utf8_roles/"$role"/default
+done
+
 rm -rf $RPM_BUILD_ROOT 
 sh -x $RPM_SOURCE_DIR/sort_patterns 11.1 1 $1 $3
 cd ..
